Add more events (#23)
Fixes #20 

We can't add an event for scoreboard changes. The contract is not
stateful with regard to the scoreboard and shouldn't be. If this should
be accomplished the Dapp needs to implement it by polling for a specific
game it is watching.

Also changes `(u32, u32)` to a `struct Field { x, y }` everywhere. They
are SCALE compatible so ABIs won't break. Could still break the
frontend. @DoubleOTheven Can you check if your Dapp still works with
this new contract?