Change scoring (#35)
Previous scoring formula that assigned bonus points based on gas usage
was confusing. Now:

- One point per painted field.
- There is a overall gas limit a contract can use.
- Specified per round on game construction. The overall budget for all
rounds per player is derived from that parameter.

Can be set to some fraction of the gas consumed of the basic player per
round. Only a fraction so that not every block a turn must be submitted
in order to win. Try `500_000_000` (500 us). The less rounds you play
the closer it needs to be to the actual usage to allow for actual turns.

For the UI nothing much changes. I recommend displaying gas left instead
of gas used per player in the player list. Can be calculated by querying
the new `gas_limit` function of the contract and subtracting the used
gas of the player from it.